https://www.reddit.com/r/MachineLearning/comments/a6zyfp/d_does_the_reparameterization_gradient_require/
"I am trying to wrap my mind around probabilistic programs and before I can understand [variational inference](https://youtu.be/ogdv_6dbvVQ?t=3160) there is something I need to make sure. I've never been aware of this before, but now I am dimly aware that there is a difference between probability being an output of a deep net - such as when using the softmax or the sigmoid activation, and well, the stuff you could find in a [probabilistic programming book](http://probmods.org/).

I am wondering how it would be possible to take the gradient of a trace of a probabilistic program even if all the individual operations are differentiable?

Even though a probabilistic program is a distribution and might be continuous and supposedly differentiable, the sampled values are not differentiable and neither is the probability of the sampled value. Factor statements are almost like a cost function, but not quite.

Is there some trick I am missing, or is WebPPL just using the score function estimator in order to do variational inference under the hood?"

I am guessing that not all continuous models are differentiable is the answer I will get. My mind is blown.