"3:15pm. The plan is starting to come together. I understand VI now. It really would not be a stretch to put in weight priors - proper ones in addition to the weight sampling that I have now. And since I am doing weight sampling it might be good to do it in parallel. Sample a batch of weights and use the strided matrix multiply and various other kinds of operations of such sort on them.
Also I should take a clue from OpenAI and the failed hypergradient experiment and use larger batch sizes than 1. If the gradients are noisy I should do it.

There is only one thing that remains to be done though. One thing I do not understand is the whole unsupervised learning connection. Why does amortized VI require the autoencoder part?

Yeah, I can understand that with weight priors if I do subsampling the thing will forget due to them, but what is the answer then?

3:25pm. I need to figure that out. Once I understand how to do amortized inference, I will have some sort of basis to make the next step.

Though this recent development of not all continuous models being differentiable is a surprise to me. I thought I would be working on a PP library, but this changes everything.

Let me do more studying. Amortized VI is next on the list."