"3:35pm. https://youtu.be/ogdv_6dbvVQ?t=3372
Let me focus on this for a bit and then I'll read the paper by Hoffman. Damn I am really confused as to when all of these parameters are suppsoed to be.

"The variational parameters that are supposed to be once per each datapoint are bound to a function that are shared across datapoints."

Wait, what?

The variational parameters in standard VI are supposed to be unique for each datapoint? That makes very little sense to me.

3:45pm. https://youtu.be/ogdv_6dbvVQ?t=3484

This particular slide is the one to figure out. I do not understand it at the moment.

3:55pm. Hmmm...actually, I am not sure I understand the VAE. I thought I did, but just how is that thing a differentiable model? Maybe it is not, but some trick is making it differentiable?

Let me read the SVI paper next."