"4:15pm. Actually, forget the paper. It is complicated. I think the reason why I do not understand it is because SVI there is intended for specific models. What about black box VI?
Let me watch that lecture by Blei on the PROBPROG conference. Maybe I am ready for it. After that, it might be good if I reread the BBB paepr and looked up stuff on the VAE.

https://youtu.be/-H2N4tVDK7I

Here is what I meant. I also saw a BBVI for deep learning video while doing the search. I'll go for it after I am done with this."