"10:25am. Actually, this paper he recommended to me is quite illuminating.
The first 3 pages make quite a lot of sense so far.

11:15am. This paper he gave me explains everything. I should not have bothered writing the reply.

11:50am. Gave a reply. Let me have breakfast. I need to study proposal distributions next.

1:15pm. Done with breakfast.

Where was I? I feel kind of restless.

I understand that proposal distributions can be used in place of prior to sample from. I assume that the log probability of a sample is then weighted according to the prior.

I am wondering what the effect of optimizing `-loq q(z,v)` is? Since it is described as entropy, I'd guess that it would push the sampled parameter `z` towards the point of greatest probability. Suppose `q` is a Gaussian distribution with parameters `mu` and `sigma`. Would the effect of minimizing `-log q` push `mu` and `sigma` towards zero?

No that is wrong. Suppose `z` and `mu` are constant. Then `sigma` woould be moved towards whatever is the most probability, meaning towards zero in order to widen the distribution. If `z` and `mu` are even however, then `sigma` would be optimized towards infinity in order to narrow the distribution.

If `sigma` was constant then `z` and `mu` would just move towards each other. I see.

1:35pm. I am thinking. Does this kind of optimization process really lead it to converging to the approximate posterior? Would it really not collapse towards infinity?

1:40pm. Ah, no wait, I got something wrong. When I described above would be minimizing `log q`. That would in fact push `z` and `mu` closer to each other and then `sigma` towards infinity. `-log q` would do exactly the opposite, it would be widening the scale of the uncertainty and the sample point and the mean of the distribution away from each other.

1:45pm. Now that makes sense. And thanks to that paper that has been linked to me, I understand how I could combine VI with deep RL. Even with a score function estimator used in stochastic nodes, the proposal distributions can still be passed the gradients from above. And the entropy bonus can be tacked to the parameters directly. It is no problem.

1:50pm. I now more or less understand both variatinal inference and how to make a probabilistic programming library. All the points of confusion have been cleared for me. I should put in some more effort in going through the lectures on them so I can understand how it is derived.

How about I do that and I will take a look at the papers I've noted yesterday."