"2:10pm. https://arxiv.org/abs/1401.0118 Black Box Variational Inference
I do not think it shows exactly how it is derived from first steps. But now that I understand how it works, it does make it much more likely that noisy KFAC is in fact variational inference. It has all the aspects of it that is for certain. If I think about what VI does, then it is pretty much exactly the same thing as noisy optimizers.

Let me read this paper.

2:20pm. /u/pyretik. I should remember this guy's name since he did help me out. It would be an insult to keep calling him 'that guy'.

2:25pm. Agh, I am having trouble figuring out where I saw the derivation. Maybe in the RL course? Bayes school?

You know, forget it. It is not that important how the ELBO...

http://users.umiacs.umd.edu/~xyang35/files/understanding-variational-lower.pdf

Let me go for this.

2:40pm. That was a good paper. I understand this now. The derivation is extremely straightforward.

2:45pm. Now let me go for the SGD as VI papers. Actually, let me step here just for a bit."