"9:20am. Actually, this paper he linked to me is pretty interesting.
```
Lemma 1 is somewhat contrived, but highlights a clear shortcoming of dropout as posterior
sampling: the dropout rate does not depend on the data. Lemma 1 means no agent employing
dropout for posterior approximation can tell the dierence between observing a set of data
once and observing it N ∫ 1 times. This can lead to arbitrarily poor decision making, even
when combined with an ecient strategy for exploration [45].
```

...

```
To understand the significance of Lemma 2, imagine a deterministic system that transitions
from s1 to s2 without reward. Suppose an agent is able to correctly quantify their posterior
uncertainty for the value V (s2)=Y ≥N(µY ,‡2
Y ). Training V (s1)=X according to (2) will
lead to zero uncertainty estimates at s1, when in fact V (s1)≥N(µY ,‡2
Y ). This observation
may appear simplistic, and may not say much more than ‘do not use the squared loss’ for
VI in this setting. However, despite this clear failing (2) is precisely the loss used by the
majority of approaches to VI for RL [17, 35, 65, 69, 20]. Note that this failure occurs even
without decision making, function approximation and even when the true posterior lies
within the variational class.
```

10:10am. This paper is amazing actually. I am going to be implementing it.

At first glance I was surprised that there is a difference between `W` and `W = W_init + W_prior`, but if I think about what happens during optimization it makes sense. The difference is essentially between `W = W_init + W_prior` and `W = W_init + grad_block W_prior`. I would not have thought to do this kind of blocking. This might just be the edge I was looking for.

11:05am. https://www.reddit.com/r/reinforcementlearning/comments/9z2b0t/r_practical_bayesian_learning_of_neural_networks/ec24181/?context=3

/u/tihokan: "Thanks for the follow-up! I do agree that uncertainty estimation has to play a very important role in RL, and current methods are not very satisfactory. I just finished reading [Randomized Prior Functions for Deep Reinforcement Learning](https://papers.nips.cc/paper/8080-randomized-prior-functions-for-deep-reinforcement-learning) and although it's definitely very interesting work, the end result still somehow feels like a big hack."

/u/abstractcontrol: "I think I'll try implementing this paper so thank you for linking it. I haven't seen it before, but I did see the previous bootstrap DQN one and did not think it had enough potential to bother with. This one does.

Reading this paper, at first glance I was surprised that there is a difference between `W` and `W = W_init + W_prior`, but if I think about what happens during optimization it makes sense. The difference is essentially between `W = W_init + W_prior` and `W = W_init + grad_block W_prior`. I would not have thought to do this kind of blocking. This might just be the edge I was looking for. At the present time it certainly seems easier than figuring out how to do variational information maximization which the noisy KFAC paper does.

Lately, I've been trying to put the failure of weight sampling in context and while studying variational inference found some papers essentially arguing that SGD itself is VI. In [the one](https://arxiv.org/abs/1710.11029) I posted recently on the ML sub, they argue that SGD does implicit VI under a different cost function than the one it is being trained on and as a consequence there is conflict with doing explicit VI which does poorly and is similar to weight sampling I am doing. In [the other](https://arxiv.org/abs/1704.04289) they actually use SGD with constant learning rates to sample from a posterior. SGD already doing VI might be a reason why explicit VI is not helping much. For that reason I'll remove weight sampling when I get back to it.

And for once I think I'll take a note from OpenAI and increase the batch size - the fact that the hypergradient descent failed is proof enough that the problem I am trying to solve (toy poker game) has some insane noise in it.

I do agree with you about this being a hack, but the real problem is not that, but that we do not really understand what it is the 'hack' of. If that were known, having the greater model would significantly improve the search. As we go along, I think it is important to keep in mind that given any moment in time the pieces necessary to make the breakthrough are there, they just need to be found and put together. To those seeking the next step that fact is hope. The No Free Lunch theorem rather than being used as an defeatist argument that all models are somehow equivalent would be better reinterpreted that no model is really useless. Meaning, that completely writing anything off should be considered a mistake. I see that as the correct attitude to take when algorithms, models or anything related to programming is concerned.

Even though I do not consider myself a researcher, I still have the responsibility of assigning credit properly to successes and failures. That is a fairly major part of being a programmer."

11:10am. How about I take a break here? I skimmed the WD paper from Grosse's group and I am amazed at how much it improves KFAC for those guys. It just makes it unstable when I'd tried it. Isn't WD just the L2 norm of the weights though? I do not get it. No wait, was L2 actually supposed to be a norm on all the weights rather than the individual ones? Huh really?

I'll check it out later. I have a bunch of things I want to read today after which I will dive into the PP book."