Let me finally resume. I'll start with the weight decay paper and work my way from there.
https://arxiv.org/abs/1810.12281
Three Mechanisms of Weight Decay Regularization

https://arxiv.org/abs/1711.05101
Fixing Weight Decay Regularization in Adam

Ah, weight decay is L2 norm without the multiplication by the learning rate. I think I multiplied it by the learning rate when I'd tried it last. So this is saying I should not do it? It might be worth a try.