"2pm. https://www.youtube.com/results?search_query=bootstrap+ml
There is a bunch of them here.

2:40pm. "**Edit**: Let me try again, there is no way it would work like this. There really cannot be a difference due to gradient blocking, sorry about this nonsense. Rather the prior should be non-linear, something like a two layer (or more) subnet in the final layer. This is actually extremely similar to [random distillation](https://arxiv.org/abs/1810.12894) as it acts as a memorization error tacked to the value prediction. If you make the output `y` of the prior `abs y`, then it does act like intrinsic reward for unknown states in a DQN."

Changed my post a little. Yeah, let me go at it from this angle instead.

2:45pm. https://youtu.be/J6I0GXyFaUk?t=16

Let me watch the videos on the bootstrap with prior.

https://youtu.be/J6I0GXyFaUk

Oh, this is pretty interesting. This could work.

3:05pm. https://colab.research.google.com/drive/1hOZeHjSdoel_-UoeLfg1aRC4uV_6gX1g#scrollTo=Gni5KBXZx5vR

Oh, here is a notebook that explains all of this.

3:15pm. I am getting super ultra distracted by the prior function stuff.

3:20pm. Ah, fuck the papers. I can't focus.

https://www.youtube.com/watch?v=ZCXg64l9R_4

Let me watch this.

3:30pm. https://youtu.be/9saauSBgmcQ
Yee Whye Teh: On Bayesian Deep Learning and Deep Bayesian Learning (NIPS 2017 Keynote)

Google recommended this to me. I haven't seen this before.

Let me stop here for a moment.

4:30pm. Done with chores. I am listening to the talk right now.

4:45pm. https://youtu.be/9saauSBgmcQ?t=1611

Reparameterization trick for discrete variables? Hmmm, this is interesting.

5pm. I've been distracted like a donkey today, especially by the randomized prior functions paper. It really fired up my imagination. Yeah, once I get back into programming, this is the thing I will be implementing. It will serve me well as an exploration method. While I can expect Bayesian RL in the near term, at least getting a boost to exploration will significantly expand the usability of current techniques.

Forget the papers by Wood. Today for a bit longer and tomorrow I will focus on the PP book."