I am up. Yesterday after finishing work I really got hopeful and once I do that my focus slips and I start enjoying things other than studying. I think I read like half of Solo Leveling. It has been a while since I binged like that. My sleep gets better when I am not under strain too.
A week from now, I think I'll take a break for a bit to compose the yearly review and the monthly PL thread post. That will take me a bit of time - a day or two. Damn, it is already this far into the year. I still have not accomplished anything of note yet. Well, hopefully with a bit more effort things will go through.

Today I will dedicate my time to the two books, but before that let me manga for a bit. After that I'll start. I am still a bit groggy.