Let me finally resume the studies. I really did fall in love with Solo Leveling. When these kinds of things start getting social they start to suck, but so far it has just been the MC fighting. It is like the good parts of Kumo when she was alone in the underground fighting against monsters.
Thankfully I've caught up and won't have to be distracted so much. Let me see if I can learn something before the day is over.

https://agentmodels.org/
Modeling Agents with Probabilistic Programs

Here is the book. I know I said I would work on a PP library until something comes along, and it did come along.

Ultimately. even if my intrinsic motivation scheme fails, I'll be able to use randomized prior functions at least.

And beyond that, even if everything works, I will still be in dead end.

I will be in a similar situation I am now and will have to frantically scramble for the next big break. I will be back to studying Bayesian ML and trying to figure out how to make true hierarchical learning work.

Nonetheless, getting intrinsically motivated exploration to work even with the current crappy methods will be a great boon.

It means I will be able to draw out the full power from the current batch of NNs and maybe finally put them to work on something useful.

Causing the Singularity is important, but so is not ending up on the streets. I have a long horizon, but it is not infinite. I have to produce results at some point - if there is any chance of doing so I have to take it.

So I will return to trying to make deep learning work. For now, let me finish my studies.