In contrast to the frenzied pace of the past days where it felt like I was drinking from the fountain of understanding today was a waste of time. Sure, there was that remark connecting anchored ensembles and randomized prior functions by /u/tihokan and I figured out the time inconsistency bias, but that is not a big deal.
Well, enough way, whether it is anchored ensembles or randomized prior functions, I will get intrinsic motivation out of the way and go to the next step. Tomorrow I think I will start by making a Gridworld environment and see how the agent does on it.

That will be my focus for the next few weeks and hopefully shorter. I'll find something that works on it with feedforward nets and then simply move from there.

I'll leave the Probabilistic project as it is right now and won't bother implementing SMC as I understand how to do it and yet have no need for it. Making deep RL work not completely horribly would be great and takes priority over anything else.

Having something that works on simple games would take a lot of pressure off me. I intend to crank up the heat until I get this to coalesce into something.