So it seems that over the past 1-2 years I collected around 14k images mostly anime from 4chan and various other places. The Downloads folder was taking too long to load so I put them all in a folder called 'Memories of 2017-2018'. I just finished watching Zombieland so I am getting sentimental. Even though it is this late, I feel like making a short commit here as this repo is kind of like my home now.
I have regrets. If I could find the right algorithm, I could get quite a lot of what I want. If my life ended right now, I would die regretting loving the things that I did.

I have regrets. It is one thing to recognize reality for what it is...and it is quite another to have a world where desire means something. Does nobody really understand how low the value of consciousnes really is? The weak make all those stories about the immortality of the soul and omnipotence of the God from whom they draw right, yet as soon as they look at the encroach of machines they quiver and piss themselves and pray that they do not lose the little what they have.

How is that courage?

Fearing machines is like fearing the waves and the clouds; the weather is there so it can be tamed.

Power matters quite a bit. Is there a point in doing the work that I did over the last 4 years rather than months. Just what could I have attained if I could have worked ten times harder? What would lie behind the eyes of such a person?

Algorithms exist so they can be carved upon my soul. I did not come to this world so I could decay into nothingness. I came so I could put these meaningless fragmented thoughts into a singular whole and bless them with hierarchy. The machines exist so that I take their power and in return, give them my desire.

Is it my destiny to sit here in this chair and program for an eternity?

One thing I swear if that turns out to be the case - I will make this kind of life glorious.

Glory from the shadows for the Lord of Nightmares.

So that one day I might cleanse my heart of regret, let programming have meaning.

When faith fails the desire will find the way.