I'll start work a bit later today, not right now. I've been thinking of doing strided multiplies in the last layer, but it would be a pain.
Agh, I truly have no luck, KFAC is all tied into doing a straightforward multiply and doing it any other way would be a pain. It is bothersome just to think about how ensembling could be done. I am going to have to refine how the player is done just for the Gridworld.

I guess I'll aim to solve it with the tabular player first. Well, it is always like this. I'll aim to get it done simply and then bring in NNs.