I am almost ready to start. Indeed, my progress for the past 6 months has been pitiful, but is more like a hangup.
Because I could not solve the super basic issue of doing good exploration with deep nets, all my progress has been essentially bottlenecked by it. I need to get it out of the way somehow.

Right now the tests are in awful shape. A major issue is that I have all these different ways of using the library. `learning1` and `learning2` each have their own custom runners and so does `poker3`.

I need to unify all of this into a common interface. Runners definitely are not trivial to create so they should be abstracted.

I sometimes think that maybe it would be better if I gone the route of not doing everything myself and use PyTorch for example, but there is no way to get good experience like that. If you are constantly stitching pieces the others have made, then you simply become that kind of programmer in the end.

If something is important then just do it yourself. If it is not, then the others can take care of it.

Power users know how the tools they are using work. That was the conclussion of the Intro to PP book and I agree with it.

One day programming languages will be used to edit and create minds - doing that with some downloaded library is beyond ridiculous. In the end, going all the way will be needed.

Let me put that aside.

Right now even though I should be starting with this, I am really not in the right mindset for it.

I know how to do it, but it is just that I have not yet build up the desire for a lengthy project such as this. Let's not forget that I threw in the towel on deep RL a few weeks ago.

I kind of need some time to switch gears and build up in my imagination how great it will be once I am done with this. I need to believe that I am on the right track and that this time this really will give me a major gain in capability.

I need to feel out the steps in my mind.

https://epubs.siam.org/doi/abs/10.1137/140964023
Randomize-Then-Optimize: A Method for Sampling from Posterior Distributions in Nonlinear Inverse Problems

Let me just study for a bit more.

I am converging to the conclusion that Bayesian inference is a scan while regular optimization is a fold operation. Ensembling is something in between.

I need to warm myself up to it more.

Once I am done with the paper, I think I will just step away from the screen for a while and sleep on it until boredom itself pushes me back into programming.

It is interesting how that works - it is not actually possible for a man to spend all his time in bed. The urge to get active will eventually take over, much like hunger and thirst.