"12:55pm. https://youtu.be/3RVGrz7MjMg Alex Graves and Marc'Aurelio Ranzato: Unsupervised Deep Learning (NeurIPS 2018 Tutorial)
This might be interesting.

https://www.youtube.com/watch?v=UwKdNzb_Q6U
[NeurIPS 2018] Loss Surfaces, Mode Connectivity, and Fast Ensembling of DNNs

But before that let me check this paper out. The randomize then optimize paper is too complicated for me to get anything out of it.

https://arxiv.org/abs/1802.10026
Loss Surfaces, Mode Connectivity, and Fast Ensembling of DNNs

The abstract for this is interesting. Training ensembles in time it takes to train a single model?

https://www.youtube.com/watch?v=0HXpnG_WnlI
Scalable Bayesian Inference - NeurIPS 2018

Meh, forget paper, the results are not that much. Let met check out this talk.

But I was not intending to do this, rather I wanted to learn more about ensembling.

https://www.youtube.com/watch?v=GU2sQgLrTx8
Loss Surfaces, fast ensembling and weight averaging of DNNs, Timur Garipov

Here is a much longer talk by Timur.

https://www.youtube.com/watch?v=Un9zObFjBH0

There is a ML for Trading course now?

1:10pm. https://www.youtube.com/watch?v=JdNohTjJKNU
Improving Deep Learning Models - Data Imbalance & Ensembling

Let me take a look at this.

1:20pm. https://youtu.be/JdNohTjJKNU?t=829

This is interesting.

1:25pm. Let me watch the Bayesian inference lecture. It is 2h. If I watch this I'll definitely not get any programming done in the day, but let me just go for it.

I have a small point of confusion regarding the whole ensembling thing. I was thinking of bootstrapping the hidden layers, but then realized that this would just be dropout. Then what do I do?

Do I do bootstrapping in just the final layer? Do I ensemble it?

What is the right thing to do in a deep net? What exactly do I want to do?

Do I just want to bootstrap the final layer? Or do I want to go with my idea? On the basis of what should I make my decision?"