"2:35pm. https://arxiv.org/abs/1703.07608 Deep Exploration via Randomized Value Functions
I never read this paper by Osband. I'll do it later. Let me resume the talk. I am really agonizing on how to do this, but I'll do it eventually.

2:50pm. https://arxiv.org/abs/1112.5016
A Scalable Bootstrap for Massive Data

This paper was mentioned in the talk.

https://arxiv.org/abs/1605.04029
Simple, Scalable and Accurate Posterior Interval Estimation

This is the paper on MCMC which he compared it to.

https://youtu.be/0HXpnG_WnlI?t=3136
"WASP/PIE is **much** faster than MCMC & highly accurate"

3pm. https://youtu.be/0HXpnG_WnlI?t=3387
"In standard Bayesian inference it is assumed that the model is correct."
"Small violations in the assumptions **sometimes** have large impacts, particularly in large datasets."

3:25pm. https://youtu.be/0HXpnG_WnlI?t=4338

Let me stop here so I can do the chores.

4:30pm. The reason it took so long because I fell long into thought while I was in the bathroom.

As expected, I really did need a bit to figure out what exactly do I want and what a principled way of doing ensembling is. I am actually coming around to a different idea compared to what I had originally. Originally I intended to sum up the values and pass them into the cost individually, but now I am thinking of reducing the ensemble values to their mean and variance and then using that to sample a new value from a Gaussian.

That will actually lead to a reduction in model error which is what I was aiming for by passing all the values into the cost individually, but will be a lot more principled as a way of tallying up the values.

4:35pm. The plan is coming together. I won't be able to avoid the need to roll up my sleeves and do some work - ensembling will require a significant extension to the library, but it won't be too hard either. The most important thing is that I get the feeling that I am doing the right thing.

And compared to the ideas I had at the start of the day, using ensembling in order to form a proposal distribution for outputs certainly meets my criteria.

I think before I start, I am going to do some more studying on Bayesian inference in the context of SMC being used to make proposal distributions for MCMC. I've heard this mentioned, but I did not look into it too much.

Of course all this will be done together with anchored weight decay."