"5:10pm. https://youtu.be/0HXpnG_WnlI?t=5899 "Highly-creative individuals have significantly > inter-hemispheric connections" (slide) "...It was an interesting side comment, we had a press release on this result and this left brain/right brain hypothesis, this whole thing where left brain people are more mathematical and the right brain people are more creative or whatever, the guy who came up with this - he is deceased now, but his son contacted us as said, well actually his dad ended up not believing the right brain/left brain thing at all and what he ended up believing was much more consistent with our thing which was - it is not right brain/left brain it is more interconnected brains that are more creative. The left brain/right brain thing is just bogus."
5:35pm. Finally done with the talk. I am thinking of posting this on the ML sub. Oh, let me just do it.

5:45pm. https://www.reddit.com/r/MachineLearning/comments/a8bw9x/d_scalable_bayesian_inference_neurips_2018/

"Video: https://youtu.be/0HXpnG_WnlI

Overall, David Dunson came off rather optimistic on the prospects of scaling up MCMC. Here is an interesting tidbit about brain regions that is rather long into the talk.

https://youtu.be/0HXpnG_WnlI?t=5899
"Highly-creative individuals have significantly > inter-hemispheric connections" (slide)
"...It was an interesting side comment, we had a press release on this result and this left brain/right brain hypothesis, this whole thing where left brain people are more mathematical and the right brain people are more creative or whatever, the guy who came up with this - he is deceased now, but his son contacted us as said, well actually his dad ended up not believing the right brain/left brain thing at all and what he ended up believing was much more consistent with our thing which was - it is not right brain/left brain it is more interconnected brains that are more creative. The left brain/right brain thing is just bogus."

https://youtu.be/0HXpnG_WnlI?t=3387
"In standard Bayesian inference it is assumed that the model is correct." (slide)
"Small violations in the assumptions **sometimes** have large impacts, particularly in large datasets." (slide)

Apart from scalability he talked a lot about the need to improve robustness for large datasets. I was not aware that this is a problem so I am highlighting this part. I'd be interested to know whether the same applies for deep learning.

Currently, I am interested in ensembling and there was a tad on it.

https://youtu.be/0HXpnG_WnlI?t=3136
"WASP/PIE is **much** faster than MCMC & highly accurate" (slide)

https://arxiv.org/abs/1605.04029
Simple, Scalable and Accurate Posterior Interval Estimation

This is the paper which he compared to MCMC.

https://arxiv.org/abs/1112.5016
A Scalable Bootstrap for Massive Data

This is the 'bag of little bootstraps' paper that was mentioned in the talk."

6pm. Posted it.

Damn, it is late. I wanted to read the Randomized Value Function paper by Osband. I'll read it for a bit and then call it a day here. I'll study it properly tomorrow along with the two papers that I highlighted in the post. Apart from that I really want to study SMC + MCMC combination methods. I want to figure out how SMC can be used to make good proposal distributions. I have a few ideas of where I could start. That paper by Wood on 'interacting particle MC' and 'SMC with rejuvenation' mentioned in the DIPPL book. There was more, but I can't recall where I'd seen it, I'll sure find something out.

Let me do a bit more and then I am done."