"9am. https://arxiv.org/abs/1806.07366v3 Neural Ordinary Differential Equations
I'll put this into the Read Later folder.

9:15am. https://www.reddit.com/r/MachineLearning/comments/a7yjkv/d_what_do_you_feel_were_the_biggest_advancements/ecb30j2/?context=3&st=jpz7l9vo&sh=259ef820

/u/progfu: "Any idea what actually happened with RUDDER? I remember when it came out people were like "omg RL is broken", but I haven't heard about it since."

/u/abstractcontrol: "I put a decent bit of thought into it months ago and realized a few things.

1) Though reward redistribution would indeed make things much better, there is still an issue of training the recurrent critic and there is not an advance there as far as I can tell. They use plain MC there.

2) At first I thought that using contribution analysis method like integrated gradients was the magic ingredient when it comes to reward redistribution, but if anything that would make things more difficult compared to using TD error as reward for the actor. The authors mention various hacks they needed to make so IG stays stable.

3) The authors did not do an ablation analysis of their various contributions so it might be that their much improved performance comes from the novel way they trained the critic which was off policy and with experience replay. Using an LSTM as the critic was also novel at the time, later [the R2D2 paper](https://openreview.net/forum?id=r1lyTjAqYX) showed significant gains from using an LSTM to do Q learning.

All in all, once I planned it out I saw that doing reward redistribution is rather complicated and now see Rudder as more of a theory paper. I admit I was one of the people really excited about it before I realized the above points.

I'd still be interested in seeing a study of its various pieces though even if I am not interested in implementing it."

9:30am. Ah, almost forgot about Satanophany, but forget it. I'll save it for the breakfast break.

Let me start reading.

How about I start with posterior sampling for RL?

I'll start with the article and then work my way to the papers.

https://ceit.aut.ac.ir/~shiry/lecture/machine-learning/papers/BRL-2000.pdf
A Bayesian Framework for Reinforcement Learning

Here is the paper where the algorithm was introduced. At this point I still do not understand what it is."