"1:40pm. Ok, enough of rest. Let me go through the bootstrap papers.
2pm. Actually the BLB algorithm is confusing to me. I am not sure what it is supposed to be doing. Well, let me just read it and then I'll go to the next thing.

2:10pm. https://analytixon.com/2016/03/14/if-you-did-not-already-know-bag-of-little-bootstraps-blb/

I need an easier explanation of it.

2:20pm. Ok, I see. So it is bootstrapping the bootstrap and then recombining. So nested sampling in other words?

To be honest, BAB is not exactly what I am looking for and it just opens up more questions for me.

2:25pm. https://www.scotthyoung.com/blog/2017/03/16/how-einstein-learned-physics/

I am getting distracted with this article. Let me finish it and then I will read the other paper.

2:35pm. Forget these papers. I definitely had enough. There is no way I am reading the 50 pages of the PIE paper.

If I need something that would be to find out what my ensembling idea that reduces a value to its mean and variance and then samples from it as a Gaussian is. Is that a good idea to do in a NN or not?

Before I go back to studying papers by Osband, I want to study SMC + MCMC combinations. I also want to look up ensembling rather than bootstrapping. It does not sit well with me that the method that gets the best performance in the RL both purposely not trains a net and ignores part of the data in the head. I dislike that."