"3:40pm. Done with chores. Let me resume.
"Our approach is inspired by [6], which applies a similar concept to the more complex context of deep reinforcement learning, but without any theoretical analysis. The essential idea is to maintain and incrementally update an ensemble of statistically plausible models, and to sample uniformly from this set in each time period as an approximation to sampling from the posterior distribution. "

The paper they are refering to the bootstrapped DQN.

4:05pm. No, this paper is not what I was looking for.

http://www.cs.colorado.edu/~mozer/Teaching/syllabi/6622/papers/Dietterich2002.pdf
https://www.analyticsvidhya.com/blog/2018/06/comprehensive-guide-for-ensemble-models/
https://arxiv.org/abs/1605.05697
https://papers.nips.cc/paper/3940-gaussian-sampling-by-local-perturbations

4:15pm. No, I can't find what I want. Ironically, the closest I've found is in the other achored ensemble paper. Maybe I should just ask Tim Pearce about whether he backpropagated through the sampled value.

Let me take a look at the paper again. After I am done with that I'll study the papers by Osband for the remainder of the day.

4:20pm. No, each NN in achored ensemble for RL paper is trained independently.

https://arxiv.org/abs/1805.11324
Bayesian Inference with Anchored Ensembles of Neural Networks, and Application to Exploration in Reinforcement Learning

This is the paper I am talking about. Let me just fire a mail and then I'll move to studying the papers by Osband including the references.

5pm. "I have an idea for which I cannot find a reference, yet it is quite simple and would like to ask for your opinion on it. From my perusal of the literature I see that ensembling methods can be put into two camps:

1) Averaging. Here every model in the ensemble is trained on all the data. This is the approach the anchored ensemble paper takes.

2) Bootstrapping. Randomized prior functions paper and bootstrapped DQN take this approach.

I'd actually like to take the first approach, but I am worried about the comment in the RPF paper that averaging worked worse for them. Nonetheless, I find purposely training on separate chunks of data inelegant and having to pass the data through a fixed random network parallel to the one being trained like in RPF and random distillation even more so.

Averaging on the other hand has a slight incoherency in that all the network in the ensemble are trained separately. I know that this is the point of ensembling, but it conceptually runs into an issue if one watch to train an actor-critic with it.

Unlike with an off-policy Q network, training the ensembles in isolation is not an option because they have to be constrained to an unified policy. If you imagine all the networks in an ensemble as threads, then they by necessity have to link before the cost function at the end.

The way to achieve that linkage is simple - take the unlinked output of the ensemble and reduce it along the outermost dimension to its mean and variance. Then using that mean and variance, sample from a Gaussian to get the linked output.

What this would do over the process of optimization is force the variance and hence the model error to zero while preventing the models from cooperating like would've been the case if I just straight up used the mean in isolation.

Using the mean and variance of outputs is done on page 3 of the anchored ensemble for RL paper, but it is done outside the model and not backproped through.

I want to try this as a part of the model and would prefer for it to work as it would be the most elegant design wise, but I am not sure what to think about adding noise to the outputs. Have you tried this or do you have any references about this kind of ensemble linking via the reparameterization trick?"

5:15pm. https://www.semanticscholar.org/paper/Risk-versus-Uncertainty-in-Deep-Learning-%3A-Bayes-%2C-Osband/dde4b95be20a160253a6cc9ecd75492a13d60c10
Risk versus Uncertainty in Deep Learning - Bayes, Bootstrap and the Dangers of Dropout

Here is a paper by Osband.

https://papers.nips.cc/paper/1197-computing-with-infinite-networks.pdf
Computing with infinite networks

Here is an old paper on GPs.

https://arxiv.org/abs/1511.06314
Why M Heads are Better than One - Training a Diverse Ensemble of Deep Networks

"Convolutional Neural Networks have achieved state-of-the-art performance on a wide range of tasks. Most benchmarks are led by ensembles of these powerful learners, but ensembling is typically treated as a post-hoc procedure implemented by averaging independently trained models with model variation induced by bagging or random initialization. In this paper, we rigorously treat ensembling as a first-class problem to explicitly address the question: what are the best strategies to create an ensemble? We first compare a large number of ensembling strategies, and then propose and evaluate novel strategies, such as parameter sharing (through a new family of models we call TreeNets) as well as training under ensemble-aware and diversity-encouraging losses. We demonstrate that TreeNets can improve ensemble performance and that diverse ensembles can be trained end-to-end under a unified loss, achieving significantly higher "oracle" accuracies than classical ensembles."

I'll start with this one.

5:25pm. The abstract for M heads paper is really encouraging. Today feels a bit off the mark. I like the PSRL papers, but so far I did not really find an answer to the question I was looking for, which is how well could I expect ensemble linking via reparameterization trick to work.

Well, I said this could take a few days, so it is no big deal. Eventually I will get through all of this.

6pm. Done with lunch. Let me continue this. I'll do the M heads paper today and the rest tomorrow.

6:10pm. "To the best of our knowledge, this is the first finding to establish that random initialization may not only be sufficient but preferred over bagging for deep networks given their large parameter space and the necessity of large training data."

This is quite an instructive paper. Interesting that bagging leads to significantly worse performance than random init.

Bagging is essentially what RPF and bootstrap DQN does.

6:15pm. "We see that shared parameter networks not only retain the performance of full ensembles, but can outperform them."

This might also be a sign that ensemble might benefit from having multiple join points for deeper nets.

6:20pm. Oh, the paper tries averaging and finds it to work worse than optimizing for independent losses.

"This expression does not depend on i — gradients backpropagated into all ensemble members are identical! Due to the averaging layer, responsibility for mistakes is shared, which eliminates gradient diversity."

Yeah, I saw this one coming. It is why I suggested the reparameterization trick instead.

6:55pm. Did a short reply to a question on the RL sub. And I am done with the paper.

Let me call it a day here. I'll die if I keep going.

Tomorrow, I will dedicate my time to studying Osband's papers unless something else pops out. And after that it will be programming time. I'll kick caution to the curb and simply go with my idea assuming it will solve the exploration problem. Going at it full force is what befits me. I've gotten a bit burned because of my previous experiments, but this time I will master the flame."